backend/requirements.txt
fastapi
uvicorn[standard]
openai
langchain
python-multipart
pandas
pypdf
Pillow
opencv-python
numpy